fix bug+ oplot
git-svn-id: https://svn.ego-gw.it/svn/advsw/Omicron/trunk@58510 47e613b6-93f3-4244-b775-e06b1756610f

+++ src/Otile.cc
@@ -120,7 +120,7 @@ bool Otile::SaveTriggers(MakeTriggers *aTriggers, const double aSNRThr, const in
   }
 
   // save segments
-  aTriggers->AddSegment((double)aT0-TimeRange/2.0+(double)aLeftTimePad,(double)aT0+TimeRange/2.0-(double)aRightTimePad);
+  aTriggers->AddSegment((double)(aT0-TimeRange/2+aLeftTimePad),(double)(aT0+TimeRange/2-aRightTimePad));
   
   return true;
 }
